get transfered bytes from admin with protocol

+++ include/metrics.h
@@ -8,9 +8,9 @@ void update_metrics_end_connection();
 void update_metrics_transfered_bytes(int n);
 
 struct metrics_manager {
-int concurrent_connections;
-int number_of_connections;
-int transfered_bytes;
+    int concurrent_connections;
+    int number_of_connections;
+    unsigned long transfered_bytes;
 };
 
 #endif
